requirements.txt
streamlit==1.35.0
nltk==3.8.1
scikit-learn==1.4.2
PyPDF2==3.0.1
python-docx==1.1.0
numpy>=1.21
pandas>=1.3
matplotlib